Use image metadata if container does not exist on atomic stop
When you call atomic stop it expects existing container which does not
have to be always true. We --rm our container in label RUN for atomicapp
because it's not useful after it finishes it's work.

This commit checks if the inspect of container was successful and if
not, it uses loads image metadata and uses STOP label from it